requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
pytesseract
python-multipart
